Add development fallbacks per module to improve DX
