+++ packages/react-scripts/config/developmentBuiltinFallbacks/constants.js
@@ -1,0 +1,11 @@
+// @remove-on-eject-begin
+/**
+ * Copyright (c) 2015-present, Facebook, Inc.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+// @remove-on-eject-end
+'use strict';
+
+module.exports = require('./createDevelopmentNodeBuiltinFallback')('constants');
